'2020-6-11晚备份'

+++ src/main/java/com/yuyi/tea/controller/PayController.java
@@ -1,19 +1,28 @@
 package com.yuyi.tea.controller;
 
 import com.google.gson.Gson;
+import com.yuyi.tea.bean.BillDetail;
+import com.yuyi.tea.bean.Customer;
+import com.yuyi.tea.bean.Notification;
 import com.yuyi.tea.bean.Order;
 import com.yuyi.tea.common.Amount;
 import com.yuyi.tea.common.CodeMsg;
+import com.yuyi.tea.common.CommConstants;
 import com.yuyi.tea.common.Result;
+import com.yuyi.tea.common.utils.TimeUtil;
 import com.yuyi.tea.exception.GlobalException;
+import com.yuyi.tea.service.CompanyService;
 import com.yuyi.tea.service.CustomerService;
 import com.yuyi.tea.service.OrderService;
+import com.yuyi.tea.service.SMSService;
 import com.yuyi.tea.service.interfaces.BalanceService;
 import com.yuyi.tea.service.interfaces.CartService;
+import com.yuyi.tea.service.interfaces.NoticeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
+import java.io.IOException;
 import java.util.List;
 
 @RestController
@@ -33,6 +42,18 @@ public class PayController {
 
     @Autowired
     private CartService cartService;
+
+    @Autowired
+    private NoticeService noticeService;
+
+    @Autowired
+    private WebSocketMINAServer webSocketMINAServer;
+
+    @Autowired
+    private SMSService smsService;
+
+    @Autowired
+    private CompanyService companyService;
 
     @GetMapping("/mobile/simulatePay/{customerId}/{value}")
     public void simulatePay(@PathVariable int customerId,@PathVariable float value){
@@ -71,10 +92,27 @@ public class PayController {
     @PostMapping("/mp/simulateCharge/{customerId}/{value}")
     public Amount mpSimulatePay(@PathVariable int customerId,@PathVariable float value){
         try{
+            long timestamp = TimeUtil.getCurrentTimestamp();
+            Customer customer = customerService.getRedisCustomer(customerId);
             //添加充值记录
             balanceService.recharge(customerId,value);
+            //添加账单记录
+            float rechargeRate = companyService.getRechargeRate();
+            BillDetail billDetail=new BillDetail(timestamp,value*rechargeRate,0,CommConstants.BillDescription.CHARGE,customer);
+            customerService.saveBillDetail(billDetail);
             //改变账户余额
             Amount balance = customerService.addBalance(customerId, value);
+            //保存通知至数据库
+            Notification notification = CommConstants.Notification.CHARGE(value, timestamp,customerId);
+            noticeService.saveNotification(notification);
+            //发送短信通知
+            smsService.sendChargeSuccess(timestamp,value,customer.getContact());
+            //发送小程序通知
+            try {
+                webSocketMINAServer.sendInfo(customer.getContact());
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
             return balance;
         }catch (Exception e){
             e.printStackTrace();
@@ -83,7 +121,7 @@ public class PayController {
     }
 
     /**
-     * 小程序支付订单
+     * 小程序支付订单(无用)
      * @param customerId
      * @param orderId
      * @return
@@ -115,12 +153,13 @@ public class PayController {
     }
 
     /**
-     * 小程序支付由购物车生成的订单
+     * 小程序支付订单
      * @param order
      * @return
      */
-    @PostMapping("/mp/payCart")
-    public String payCart(@RequestBody Order order){
+    @PostMapping("/mp/payCart/{isCart}")
+    public String payCart(@RequestBody Order order,@PathVariable boolean isCart){
+        long currentTimestamp = TimeUtil.getCurrentTimestamp();
         Order currentOrder=orderService.getOrder(order.getUid());
         currentOrder.setBuyerPs(order.getBuyerPs());
         currentOrder.setDeliverMode(order.getDeliverMode());
@@ -135,10 +174,15 @@ public class PayController {
         customerService.checkBalance(ingot,credit,balance);
         //扣费
         customerService.pay(ingot,credit,currentOrder.getCustomer().getUid());
+        //添加账单记录
+        BillDetail billDetail=new BillDetail(currentTimestamp,-ingot,-credit,CommConstants.BillDescription.BUY_PRODUCT,order.getCustomer());
+        customerService.saveBillDetail(billDetail);
         //改变订单状态
         orderService.updateOrderPayed(currentOrder);
-        //删除购物车相应内容
-        cartService.deleteCartProductByOrder(currentOrder);
+        if(isCart){
+            //删除购物车相应内容
+            cartService.deleteCartProductByOrder(currentOrder);
+        }
         return "success";
     }
